// src/app/models/genes.ts
export interface IGenes {
    body: Array<number>;
    pattern: Array<number>;
    eyeColor: Array<number>;
    eyeType: Array<number>;
    bodyColor: Array<number>;
    patternColor: Array<number>;
    accentColor: Array<number>;
    wild: Array<number>;
    mouth: Array<number>;
    environment: Array<number>;
    secret: Array<number>;
    unknown: Array<number>;
}